[minor] Fixing tabs and indentation
We want to use tabs for alignment, spaces for indentation. Fixing as
much of that as I can. Also, there is some stuff which looks like it
might be screwed up, indentation-wise (especially in game.py). but I
wasn't sure and didn't fix it. Game needs major testing (make sure it
runs, for starters).